Delineate local & global styles with styleImporter
I wish I could figure out how to keep webpack from mangling the class
names for the Picnic CSS. But keep it for all of my local CSS.

If I can't, then maybe this is an acceptable workaround. I'll be able to
keep moving forward, for now.

And if I ever figure out how to prevent Picnic CSS mangling, it will be
straightforward to modify all "styles.global" occurrences.