Attempt to speed up navigation without Immutable
This thing is horribly slow. Four seconds to navigate to a detail view.
It feels so crappy and dumb to use. Not delightful at all.

I tried fixing it with an insanely aggressive `shouldComponentUpdate` in
`Calendar`. It's actually TOO aggressive, and causes bugs. But still
things are slow.

I like the way I've tidied things up, in general, even if it didn't have
the affect I desired. So I'm leaving it in.