Fix goofy route transition bug
I don't understand why, but here's what was happening:

1. Refresh the page at `/bessie-coleman`
2. Click the "a life" icon to go to the home page
3. The `@@reduxReactRouter/routerDidChange` action was fired twice
4. It seems to try to (briefly?) render a User page with a slug of `''`
5. This causes warnings or errors in the components User, Calendar,
   Week, and WeekDetail, since they expect `slug` to be present

Note that starting at `/`, navigationg to `/bessie-coleman` by clicking
the link on the home page, and then clicking the link back to the home
page – that works just fine.

The (a?) solution is to add annoying logic that safeguards against this
blank slug condition. In all of those components.

That's what I did here. I don't really like it. I know that logic is
going to annoy me (or someone) in the future, and we'll try to remove
it, and everything will seem fine until someone discovers this weird
thing lurking there again. Who knows, maybe ReactRouter or whoever's
culpable will be fixed by then. If not, hopefully I/we come across this
commit message before we go removing that logic.