Fix scroll position on new pages
At least, sorta fix it.

I got rid of the `scroll-behavior` library several commits back, because
it created a pretty terrible user experience. Without it, though, the
scroll position stays the same between all page loads. At least, as best
it can.

How this played out is that you'd scroll to the bottom of the long home
page, then click the link to "Team", and then you'd be (weirdly) at the
bottom of the Team page, instead of the top. Same when clicking through
to Bessie Coleman or Bayard Rustin.

Adding a simple `window.scrollTo(0,0)` in `componentDidMount` is enough
to fix it. I did this using a decorator, so that the logic—if it ever
needs to change—can be changed in one clean place.

I'm not sure how this will play out with other components that use
`componentDidMount`, but we'll cross that bridge when we come to it.