Add SignedInNav
Now the user can tell if they're signed in or not. And they can sign
out.

Some of the links don't work yet:

* Link to `/account`
* Link to `/quiz`

Adding those pages is coming, but first, I need to make it remember that
the user is signed in when the page first loads from the server.