Load auth state correctly on server
This required diligent cookie management. Basically, I don't have a
reliable way to set cookies on the server. So the LOGIN_SUCCESS and
LOGIN_FAIL reducers can't set cookies. They also couldn't read cookies
before, but I've solved that by passing cookies from the ApiClient into
the SUCCESS-type action creators.

So now cookies are set in the `login` and `logout` action creators, and
are never set in the reducers.

The `load` action is fired on the server, and so also can't set or read
cookies. For this reason, it passes a newfangled `onlyIf` function to
the ApiClient, which can use cookies, and will abort without making an
unnecessary request if the condition fails.

Ideally, the `LOGIN_FAIL` action handler would be able to clear cookies,
but since it's fired on the server, it can't. If a user's Google session
expires (meaning the user does not manually sign out, but instead uses
the site until Google signs them out), they'd be left with a krufty
`user` and `idToken` cluttering up their cookies, which would mean the
`load` request would keep happening.

To get around that, the `googleAuth` helper now cleans up these cookies
after it initalizes.