Improve UX of form saves & errors
The biggest change here is that the `auth` reducer is the one
responsible for updating the user. That's the way it should be! And it
resulted in the redux-form experience Just Working—meaning that, on
save, the "pristine" setting gets re-set, and I can use it to change the
text of the "Save" button. Cool.

It's definitely a better experience than the current in-production form.
It even gives useful errors for dates when in a browser that doesn't
support input[type=date] (lookin' at you, Safari).

Speaking of Safari, Safari's what necessitated the change to the Signin
styling. It was not centered on Safari.