Get single Quiz question working
It's unstyled, and the code is not very Redux-ey, but that's probably ok
for now. It works, and it will be way faster to bring in the rest of the
quiz if I don't try to rearchitect the whole thing.

Rearchitecting in this case wouldn't get me much, I don't think.