Improve detail views a bunch
With this change, you can now edit events in the month view. The display
of the date at the top has also been improved – it shows the range of
dates (correctly!) instead of saying "Month of June 15", which confused
people.

The form for months is bananas – it lists out every single date in the
four weeks! That's a ridiculous set of radio buttons. A better date
picker is coming next.

There's some weirdness when going mobile<-->notMobile, but that's an
edge case I optimized enough for now. Time to move on to more important
things!